The following syscalls should not be blocked by seccomp
Capabilities block these syscalls.

mount, umount2, unshare, reboot and name_to_handle_at are all needed to
run systemd as pid1 in a container, they work fine with sys_admin disabled
and have functionality in the kernel that is available to a non privileged
process.  There is no easy way to discover which syscalls are blocked, so
we end up more likely with the user doing a --privileged.

Signed-off-by: Dan Walsh <dwalsh@redhat.com>