Attempt to mimic changes from upstream PR#27467
Problem: if a process that's been started for 'docker exec' exits fast
enough, the daemon can receive a 'process exited' update from containerd
before it starts passing stdio data back and forth, losing its output.

Why it happens: the 'process exited' state change message is only
processed after acquiring a lock for the container, and while most of
the exec() setup is done while holding that lock, the lock is freed when
libcontainerd calls back to set up the passing of stdio data.

Proposed change: call getExecConfig() before AddProcess(), and modify
AddProcess() to take a callback that the daemon can hand it that
remembers the exec.Config value that was retrieved.

The effect should be the same as that of the changes in
https://github.com/docker/docker/pull/27467, but without as much
refactoring.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>