BACKPORT: Fix issue with cp to container volume dir
Upstream reference: https://github.com/docker/docker/pull/28047
Upstream reference: https://github.com/docker/docker/issues/27773
Upstream reference: https://github.com/projectatomic/docker/issues/207

In some cases, attempting to `docker cp` to a container's volume dir
would fail due to the volume mounts not existing after performing a
bind-mount on the container path prior to doing a pivot_root.

This does not seem to be effecting all systems, but was found to be a
problem on centos.
The solution is to use an `rbind` rather than `bind` so that any
existing mounts are carried over.

The `MakePrivate` on `path` is no longer neccessary since we are already
doing `MakeRPrivate` on `/`.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>
Signed-off-by: Antonio Murdaca <runcom@redhat.com>