redhat-ci: build and test in a fedora container
The upstream tests are being run on Debian rather than distro-specific
containers because there is no golang compiler on AH to compile
gen_dockerfile.go. We should consider porting it to e.g. python in the
future, but for now, we can work around this by building the Dockerfile
in a container.