Error out if selinux is not supported with overlay
Previously we had a check in docker which used to error out if selinux was
enabled and graph driver was overlay/overlay2. Now we have removed that
check and selinux is enabled by default in docker.

Now it is possible that users are using overlay/overlay2 but still using
old kernel which does not support selinux in overlay. In that case it is
better to error out and not start docker daemon and tell user to either
upgrade to a newer kernel which supports selinux for overlay or disable
selinux in docker. One can disable selinux in docker by specifying
--selinux-enabled=false in docker daemon options.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>