Call sd_journal_get_fd() earlier
The journal API opens the journal files when the journal handle is
opened by sd_journal_open(), and uses an inotify descriptor to notice
when files have been removed, but it doesn't set up that descriptor
until the first time that sd_journal_get_fd() is called (either by a
client, or as part of sd_journal_wait()).

We hadn't been doing that until our initial read-through of entries was
done, meaning that we've been missing file deletion events that occurred
while we were reading the journal.  Make that window shorter.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>