journal: don't stop reading early in non-follow mode
If the output channel gets stalled while reading the journal in
non-follow mode, don't stop right away.  Doing this correctly means that
we need to detect when we've hit the end of the log, so start returning
a value indicating whether or not we ran out of entries to read.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>